requirements.txt
openai == 0.28.1
Flask == 2.3.3
langchain == 0.0.303
pinecone-client == 2.2.4
tiktoken == 0.5.1